packages/rules/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crdc-precheck-rules"
version = "0.1.0"
description = "Rules DSL and evaluator for CRDC PreCheck validations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.6,<3.0",
    "PyYAML>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.3,<25.0",
    "mypy>=1.9,<2.0",
    "pytest>=8.1,<9.0",
    "ruff>=0.3,<1.0"
]

[tool.setuptools]
package-dir = {"" = "rules"}

[tool.setuptools.packages.find]
where = ["rules"]
